add `file_utils` module
